Move logging-related modules to a separate directory
