+++ apps/bot/src/logging/logger.ts
